requirements.txt
python-dotenv
colorama
ffmpeg-python
